src/services/activity-service/requirements.txt
fastapi==0.134.0
uvicorn==0.41.0
asyncpg==0.31.0
valkey-glide==2.2.7